Move extensions config to its own file
